Refactoring types
Mostly moving types -> esm-globals
