+++ packages/shell/esm-app-shell/src/helpers.ts
@@ -1,4 +1,3 @@
-import { getLoggedInUser, userHasAccess } from "@openmrs/esm-framework";
 import { LifeCycles } from "single-spa";
 
 export const emptyLifecycle: LifeCycles<never> = {
@@ -12,10 +11,6 @@ export const emptyLifecycle: LifeCycles<never> = {
     return Promise.resolve();
   },
 };
-
-export function routePrefix(prefix: string, location: Location) {
-  return location.pathname.startsWith(window.getOpenmrsSpaBase() + prefix);
-}
 
 export function routeRegex(regex: RegExp, location: Location) {
   const result = regex.test(
