+++ packages/shell/esm-app-shell/src/index.ts
@@ -9,11 +9,35 @@ import "@openmrs/esm-framework";
 import type { SpaConfig } from "@openmrs/esm-framework/src/internal";
 
 export function setupPaths(config: SpaConfig) {
+  let error = false;
+  if (!config.apiUrl) {
+    console.error(
+      "initializeSpa() was called without supplying an apiUrl. This means that the application cannot communicate with the backend."
+    );
+    error = true;
+  }
+
+  if (!config.spaPath) {
+    console.error(
+      "initializeSpa() was called without supplying a spaPath. This means that the application cannot properly generate urls."
+    );
+    error = true;
+  }
+
+  if (error) {
+    throw new Error(
+      "One or more required properties in the basic configuration of the application was missing and the application cannot be rendered. Please see the browser console for details."
+    );
+  }
+
   window.openmrsBase = config.apiUrl;
   window.spaBase = config.spaPath;
   window.spaEnv = config.env || "production";
   window.spaVersion = process.env.BUILD_VERSION;
-  window.getOpenmrsSpaBase = () => `${window.spaBase}/`;
+  const spaBaseWithSlash = window.spaBase.endsWith("/")
+    ? window.spaBase
+    : window.spaBase + "/";
+  window.getOpenmrsSpaBase = () => spaBaseWithSlash;
 }
 
 export function setupUtils() {
